MacOS support, Get issuer from server, common_name is sourced instead of stored in config
* Break out imports to clean up setting trust in MacOS
* Get issuer from server for Root CA Certificate for MacOS validation
* Add comments to explain logic around certificate validation on MacOS
and how it differs from Linux
* Fix darwin validity check, move common_name to be friendly_name
* Change validity check for root vs client on Darwin
* Move from common_name in config to friendly name
* Have friendly name get USER and HOSTNAME from system instead of having
the user define it
* Fail out validation if issuer not found
* Add support for deleting expired certs in MacOS. Move lifetime to be seconds and continue to default to 18 hours

BREAKING CHANGE: common_name is no longer in the config and is instead
sourced and generated to match the same friendly name pattern found
using certutil. This means any currently generated CSRs will use the old
style but may not actually delete properly. These may need to be deleted
manually